pyproject.toml
[project]
name = "crosswalk-v2"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "pyzmq>=26.4.0",
    "zmq>=0.0.0",
    "pydantic>=2.0.0",
    "gpiozero>=2.0.1",
    "fastapi>=0.104.0",
    "uvicorn>=0.24.0",
    "transitions>=0.9.2",
]

[dependency-groups]
dev = [
    "ipython>=9.3.0",
    "PyYAML>=6.0.0",
    "numpy>=1.24.0",
    "mutagen>=1.47.0",
    "Pillow>=10.0.0",
]
